Don't download images if we already have them
Not a very sophisticated caching strategy but it'll do for now

+++ app/image_store.rb
@@ -15,7 +15,9 @@ class ImageStore
   def download(uri)
     name = filename(uri)
 
-    File.open(name, "wb") { |file| file.print(fetcher.call(uri)) }
+    if !File.exist?(name)
+      File.open(name, "wb") { |file| file.print(fetcher.call(uri)) }
+    end
 
     ImageFile.new(name, image_dimensions_calculator: image_dimensions_calculator)
   end
